+++ previous-versions/common/RELEASES.md
@@ -721,3 +721,13 @@ Based on:
 - [typescript v0.40.1] previous-versions/common
 ### Releases
 - [NPM v0.40.1] https://www.npmjs.com/package/@codat/common/v/0.40.1 - previous-versions/common
+
+## 2023-10-03 00:03:43
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.93.1 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.40.2] previous-versions/common
+### Releases
+- [NPM v0.40.2] https://www.npmjs.com/package/@codat/common/v/0.40.2 - previous-versions/common